switched to built-in layout capability
