+++ frontend/src/app/layout.tsx
@@ -1,32 +1,43 @@
-import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
+"use client"
+import React, { ReactNode } from 'react'
+import Link from "next/link";
+import styles from "./page.module.css";
+import { usePathname, useSearchParams } from 'next/navigation'
 
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
 
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
+let sites: string[][] = [["Startseite", "/"], ["Testseite", "/test_page"]];
 
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
+function Nav_Menu() {
+  const path = usePathname();
+  let links = sites.map((site, index) => {
+    if(site[1]==path){
+      return<Link href={site[1]} key={index} className={styles.navbutt +" "+ styles.active}>{site[0]}</Link>
+    }
+    return<Link href={site[1]} key={index} className={styles.navbutt}>{site[0]}</Link>
+  });
+  return <nav key = 'navMen' className={styles.navbar}>{links}</nav>;
+}
+
 
 export default function RootLayout({
   children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
+}: {
+  children: React.ReactNode
+}) {
   return (
-    <html lang="en">
-      <body className={`${geistSans.variable} ${geistMono.variable}`}>
-        {children}
+    <html lang="de">
+      <body>
+        <div className={styles.back_layer}>
+          <Nav_Menu />
+          <div className={styles.page}>
+            <main className={styles.main}>
+              <div className={styles.content}>
+                {children}
+              </div>
+            </main>
+          </div>
+        </div>
       </body>
     </html>
-  );
+  )
 }
