+++ frontend/src/app/page.module.css
@@ -190,23 +190,20 @@ a.secondary {
   z-index: 1000; /* Keeps it above other content */
 }
 .navbutt {
-  background-color:antiquewhite;
+  background-color: antiquewhite;
   padding: 20px; /* Adds space inside the content area */
   border-radius: 10px; /* Rounds the corners slightly */
   box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1); /* Adds a subtle shadow */
   max-width: 800px; /* Limits width for better readability */
   margin: 0 auto; /* Centers content */
   line-height: 1.6; /* Improves text readability */
+  text-decoration: none;
+  font-weight: bold;
+  color: black;
 }
 
-.navbutt_active {
-  background-color:peachpuff;
-  padding: 20px; /* Adds space inside the content area */
-  border-radius: 10px; /* Rounds the corners slightly */
-  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1); /* Adds a subtle shadow */
-  max-width: 800px; /* Limits width for better readability */
-  margin: 0 auto; /* Centers content */
-  line-height: 1.6; /* Improves text readability */
+.navbutt.active {
+  background-color: peachpuff;
 }
