+++ frontend/src/app/page.tsx
@@ -7,9 +7,9 @@ import Layout from "../../generic_functions/base_layout";
 
 export default function Home() {
   return (
-    <Layout>
+    <div>
           <h1>Startseite</h1>
           <p>Nur hier um zu existieren</p>
-    </Layout>
+    </div>
   );
 }
